+++ src/interfaces/config.interface.ts
@@ -1,4 +1,5 @@
 export interface IConfig {
     NODE_ENV: string;
     APP_NAME: string;
+    BING_API: string;
 }