+++ src/interfaces/context.interface.ts
@@ -1,0 +1,4 @@
+export interface BingImageContextType {
+    backgroundImage: string;
+    fetchBackgroundImage: () => void;
+}