+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- WIP: bing api request
+- WIP: stores and context
+
 ## [0.4.0] - 2024-06-20
 
 ### Added
@@ -20,7 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Fixed
 - minor layout fixes
-
 
 ## [0.3.0] - 2024-06-20